feat: добавил логику постинга контейнеров на сайт

+++ ContainerFather.Core/UseCases/Containers/Models/GetContainerListResponse.cs
@@ -13,7 +13,7 @@ public class GetContainerListResponse
     public decimal Price { get; set; }
     public CurrencyEnum Currency { get; set; }
     public string Address { get; set; }
-    public float Latitude { get; set; }
-    public float Longitude { get; set; }
+    public double Latitude { get; set; }
+    public double Longitude { get; set; }
     public DateTimeOffset CreatedAt { get; set; }
 }
